pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinellium"
version = "0.2.0"
description = "Frida-based Android/iOS RASP auditing framework"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Sentinellium Authors"},
]
dependencies = [
    "frida>=16.0.0",
    "frida-tools>=12.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]
ebpf = [
    "bcc",
]
ci = [
    "pytest>=7.4.0",
]

[project.scripts]
sentinellium = "host.cli:cli"

[tool.setuptools.packages.find]
include = ["host*", "gadget*", "ebpf*", "ci*"]

[tool.pytest.ini_options]
testpaths = ["tests/host"]
pythonpath = ["."]
